# internal/database/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type ScanStatus string

const (
	ScanStatusQueued    ScanStatus = "queued"
	ScanStatusRunning   ScanStatus = "running"
	ScanStatusCompleted ScanStatus = "completed"
	ScanStatusFailed    ScanStatus = "failed"
	ScanStatusCancelled ScanStatus = "cancelled"
)

func (e *ScanStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = ScanStatus(s)
	case string:
		*e = ScanStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for ScanStatus: %T", src)
	}
	return nil
}

type NullScanStatus struct {
	ScanStatus ScanStatus `json:"scan_status"`
	Valid      bool       `json:"valid"` // Valid is true if ScanStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullScanStatus) Scan(value interface{}) error {
	if value == nil {
		ns.ScanStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.ScanStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullScanStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.ScanStatus), nil
}

func (e ScanStatus) Valid() bool {
	switch e {
	case ScanStatusQueued,
		ScanStatusRunning,
		ScanStatusCompleted,
		ScanStatusFailed,
		ScanStatusCancelled:
		return true
	}
	return false
}

type Scan struct {
	ID           uuid.UUID          `json:"id"`
	URL          string             `json:"url"`
	Status       ScanStatus         `json:"status"`
	ErrorMessage string             `json:"error_message"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}