# openapi/openapi.yaml
openapi: 3.1.0

info:
  title: AksesCheck ID API
  version: 0.1.0
  description: API untuk pemindaian aksesibilitas website.

servers:
  - url: http://localhost:8080/api/v1

paths:
  /health:
    get:
      operationId: getHealth
      summary: Memeriksa kondisi API
      responses:
        "200":
          description: API, PostgreSQL, dan Redis tersedia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Salah satu layanan tidak tersedia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /scans:
    post:
      operationId: createScan
      summary: Membuat pemindaian baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "201":
          description: Scan berhasil dibuat dan dimasukkan ke antrean
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: URL atau body request tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Kesalahan database atau antrean
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scans/{scanId}:
    get:
      operationId: getScan
      summary: Mengambil detail scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Detail scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Format scan ID tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Kesalahan database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - database
        - redis
      properties:
        status:
          type: string
          enum:
            - ok
            - degraded
        database:
          type: string
          enum:
            - ok
            - unavailable
        redis:
          type: string
          enum:
            - ok
            - unavailable

    CreateScanRequest:
      type: object
      additionalProperties: false
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
          examples:
            - https://example.com

    ScanStatus:
      type: string
      enum:
        - queued
        - running
        - completed
        - failed
        - cancelled

    Scan:
      type: object
      required:
        - id
        - url
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/ScanStatus"
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string